requirements.txt
requests
playsound
pyaudio
webrtcvad
pydub
tinydb
pythainlp
feedparser
pandas
speechrecognition
gtts
pydub
python-vlc
pafy
youtube-dl